media/views.py
import cv2
import tensorflow as tf
from rest_framework import viewsets
from rest_framework.response import Response

from .models import MediaFile
from .serializers import MediaFileSerializer

class MediaFileViewSet(viewsets.ModelViewSet):
    queryset = MediaFile.objects.all()
    serializer_class = MediaFileSerializer

    def predict(self, request, pk):
        media_file = self.get_object()
        model_path = media_file.ai_model.path
        model = tf.keras.models.load_model(model_path)

        # Video stream processing logic
        video_stream = cv2.VideoCapture(media_file.video.path)
        while True:
            ret, frame = video_stream.read()
            if not ret:
                break

            # Perform prediction on each frame
            # ...

        video_stream.release()

        return Response({'result': prediction})
