requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
environ==1.0
graphene==3.3
graphene-django==3.2.0
graphql-core==3.2.3
graphql-relay==3.2.0
inflection==0.5.1
mpmath==1.3.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
promise==2.3
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
setuptools==69.2.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
# pycocotools
EasyDict==1.7
opencv-python
shapely
Cython
scipy
pandas
pyyaml
json_tricks
scikit-image
yacs>=0.1.5
tensorboardX